Add support for aws_apigatewayv2_route
