+++ pkg/iac/terraform/state/test/apigatewayv2_route/terraform.tfstate
@@ -1,0 +1,40 @@
+{
+  "version": 4,
+  "terraform_version": "1.0.10",
+  "serial": 4,
+  "lineage": "test-lineage",
+  "outputs": {},
+  "resources": [
+    {
+      "mode": "managed",
+      "type": "aws_apigatewayv2_route",
+      "name": "example",
+      "provider": "provider[\"registry.terraform.io/hashicorp/aws\"]",
+      "instances": [
+        {
+          "schema_version": 0,
+          "attributes": {
+            "api_id": "abcd",
+            "api_key_required": false,
+            "authorization_scopes": null,
+            "authorization_type": "NONE",
+            "authorizer_id": "",
+            "id": "dcab",
+            "model_selection_expression": "",
+            "operation_name": "",
+            "request_models": null,
+            "request_parameter": [],
+            "route_key": "$default",
+            "route_response_selection_expression": "",
+            "target": ""
+          },
+          "sensitive_attributes": [],
+          "private": "bnVsbA==",
+          "dependencies": [
+            "aws_apigatewayv2_api.example"
+          ]
+        }
+      ]
+    }
+  ]
+}
