+++ pkg/middlewares/testdata/aws_apigatewayv2_api_body_openapiv3.yml
@@ -1,0 +1,24 @@
+openapi: "3.0.1"
+info:
+  title: "example-api"
+  version: "2021-12-07 13:55:09UTC"
+servers:
+- url: "https://abcd.execute-api.eu-west-1.amazonaws.com/{basePath}"
+  variables:
+    basePath:
+      default: ""
+paths:
+  /example:
+    get:
+      responses:
+        default:
+          description: "Default response for GET /example"
+    post:
+      responses:
+        default:
+          description: "Default response for POST /example"
+  /example2:
+    get:
+      responses:
+        default:
+          description: "Default response for GET /example2"
