+++ pkg/remote/aws/repository/apigatewayv2_repository.go
@@ -11,6 +11,7 @@ import (
 
 type ApiGatewayV2Repository interface {
 	ListAllApis() ([]*apigatewayv2.Api, error)
+	ListAllApiRoutes(apiId *string) ([]*apigatewayv2.Route, error)
 	ListAllVpcLinks() ([]*apigatewayv2.VpcLink, error)
 	ListAllApiAuthorizers(string) ([]*apigatewayv2.Authorizer, error)
 }
@@ -41,6 +42,22 @@ func (r *apigatewayv2Repository) ListAllApis() ([]*apigatewayv2.Api, error) {
 		return nil, err
 	}
 
+	r.cache.Put(cacheKey, resources.Items)
+	return resources.Items, nil
+}
+
+func (r *apigatewayv2Repository) ListAllApiRoutes(apiID *string) ([]*apigatewayv2.Route, error) {
+	cacheKey := fmt.Sprintf("apigatewayv2ListAllApiRoutes_api_%s", *apiID)
+	v := r.cache.Get(cacheKey)
+
+	if v != nil {
+		return v.([]*apigatewayv2.Route), nil
+	}
+
+	resources, err := r.client.GetRoutes(&apigatewayv2.GetRoutesInput{ApiId: apiID})
+	if err != nil {
+		return nil, err
+	}
 	r.cache.Put(cacheKey, resources.Items)
 	return resources.Items, nil
 }
