+++ pkg/resource/aws/aws_apigatewayv2_route.go
@@ -1,0 +1,3 @@
+package aws
+
+const AwsApiGatewayV2RouteResourceType = "aws_apigatewayv2_route"
