+++ pkg/resource/aws/testdata/acc/aws_apigatewayv2_route/.driftignore
@@ -1,0 +1,2 @@
+*
+!aws_apigatewayv2_route
