+++ pkg/resource/resource_types.go
@@ -139,11 +139,14 @@ var supportedTypes = map[string]ResourceTypeMeta{
 	}},
 	"aws_appautoscaling_policy":           {},
 	"aws_appautoscaling_scheduled_action": {},
-	"aws_apigatewayv2_api":                {},
-	"aws_apigatewayv2_vpc_link":           {},
-	"aws_apigatewayv2_authorizer":         {},
-	"aws_launch_template":                 {},
-	"aws_launch_configuration":            {},
+	"aws_apigatewayv2_api": {children: []ResourceType{
+		"aws_apigatewayv2_route",
+	}},
+	"aws_apigatewayv2_route":      {},
+	"aws_apigatewayv2_vpc_link":   {},
+	"aws_apigatewayv2_authorizer": {},
+	"aws_launch_template":         {},
+	"aws_launch_configuration":    {},
 
 	"github_branch_protection": {},
 	"github_membership":        {},
